+++ include/dislocker/metadata/datums.h
@@ -30,28 +30,7 @@
 #include "dislocker/metadata/guid.h"
 #include "dislocker/ntfs/clock.h"
 #include "dislocker/ntfs/encoding.h"
-
-
-
-/**
- * Cipher used within BitLocker
- */
-enum cipher_types
-{
-	STRETCH_KEY   = 0x1000,
-	AES_CCM_256_0 = 0x2000,
-	AES_CCM_256_1 = 0x2001, 
-	EXTERN_KEY    = 0x2002,
-	VMK           = 0x2003,
-	AES_CCM_256_2 = 0x2004,
-	HASH_256      = 0x2005,
-	
-	AES_128_DIFFUSER    = 0x8000,
-	AES_256_DIFFUSER    = 0x8001,
-	AES_128_NO_DIFFUSER = 0x8002,
-	AES_256_NO_DIFFUSER = 0x8003,
-};
-typedef uint16_t cipher_t;
+#include "dislocker/encryption/encommon.h"
