+++ src/dislocker-fuse.c
@@ -45,7 +45,6 @@
 
 #include "dislocker/common.h"
 #include "dislocker/dislocker.h"
-#include "dislocker/encryption/encommon.h"
 #include "dislocker/metadata/metadata.h"
