+++ src/encryption/decrypt.c
@@ -25,6 +25,7 @@
 #include "dislocker/common.h"
 #include "dislocker/encryption/diffuser.h"
 #include "dislocker/encryption/decrypt.h"
+#include "dislocker/encryption/encommon.priv.h"
